Extracts characters (player and enemy) classes.
Also decided to renmae the images module into imaging, as the images
module name may find another use in the future.
